OCR/OCR_Pro-main/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.13.25
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.11.0
altgraph==0.17
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
asn1crypto==1.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.11.1
bleach==6.1.0
bs4==0.0.2
cachelib==0.1.1
cachetools==4.1.1
certifi==2024.8.30
cffi==1.14.1
cfgv==3.2.0
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
colorama==0.4.3
configparser==7.1.0
cos-python-sdk-v5==1.7.7
crcmod==1.7
cryptography==3.0
cssselect==1.2.0
Cython==0.29.21
decorator==5.1.1
defusedxml==0.7.1
dicttoxml==1.7.4
distlib==0.3.1
distro==1.9.0
dnspython==2.0.0
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.8.0.post1
fake-useragent==1.5.1
fastjsonschema==2.20.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==5.0.0
Flask-Session==0.3.2
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
frozenlist==1.4.1
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
google-api-core==1.22.0
google-auth==1.20.0
google-cloud-core==1.4.0
googleapis-common-protos==1.52.0
greenlet==0.4.16
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
identify==1.4.25
idna==2.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
IPy==1.0
ipython==8.12.3
itsdangerous==1.1.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
jmespath==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.2.16
langchain-core==0.2.40
langchain-text-splitters==0.2.4
langsmith==0.1.120
lxml==4.5.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.1.0
mysqlclient==2.0.1
Naked==0.1.31
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ndg-httpsclient==0.5.1
news_scraper==0.4.0
nodeenv==1.4.0
numpy==1.24.4
oauthlib==3.2.2
openai==1.45.0
orjson==3.10.7
oss2==2.5.0
outcome==1.3.0.post0
packaging==24.1
pandocfilters==1.5.1
paramiko==2.7.1
parse==1.20.2
parso==0.8.4
pathtools==0.1.2
pefile==2019.4.18
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pluggy==0.13.1
pre-commit==2.6.0
prompt_toolkit==3.0.47
protobuf==3.12.4
psutil==5.7.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pure_eval==0.2.3
py==1.9.0
py7zr==0.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.8
pydantic==2.9.1
pydantic_core==2.23.3
pyee==11.1.1
Pygments==2.18.0
PyInstaller==3.6
pymemcache==3.2.0
pymongo==3.11.0
pymssql @ file:///C:/pymssql-2.1.4-cp38-cp38-win_amd64.whl
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyMySQL==0.10.0
PyNaCl==1.4.0
pyOpenSSL==19.1.0
pyotp==2.4.0
pyparsing==2.4.7
pyperclip==1.9.0
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-engineio==3.13.1
python-memcached==1.59
python-socketio==4.6.0
python3-memcached==1.51
pytz==2020.1
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==5.3.1
pyzmq==26.2.0
qiniu==7.2.8
redis==3.5.3
referencing==0.35.1
requests==2.28.1
requests-html==0.10.0
requests-oauthlib==1.3.1
rpds-py==0.20.0
rsa==4.6
schedule==1.2.2
selenium==4.24.0
shellescape==3.8.1
six==1.15.0
sniffio==1.3.1
some-package==0.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.34
stack-data==0.6.3
tenacity==8.5.0
texttable==1.6.2
tinycss2==1.3.0
toml==0.10.1
tornado==6.4.1
tox==3.18.1
tqdm==4.66.5
traitlets==5.14.3
translate==3.5.0
trio==0.26.2
trio-websocket==0.11.1
tweepy==4.12.1
typing_extensions==4.12.2
unrar==0.4
upyun==2.5.5
urllib3==1.26.20
virtualenv==20.0.29
w3lib==2.2.1
watchdog==0.10.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
Werkzeug==1.0.1
winshell==0.6
WMI==1.5.1
wsproto==1.2.0
xmltodict==0.12.0
yarg==0.1.9
yarl==1.11.1
zipp==3.20.2
zope.event==4.4
zope.interface==5.1.0
